增加一些openVINO自带的函数

+++ src/main/java/org/openvino/java/core/Input.java
@@ -2,13 +2,23 @@ package org.openvino.java.core;
 
 import org.openvino.java.base.OpenVINOCls;
 
+/**
+ * Class holding preprocessing information for one input
+ * From preprocessing pipeline perspective, each input can be represented as:
+ *  - User's input parameter info (InputInfo::tensor)
+ *  - Preprocessing steps applied to user's input (InputInfo::preprocess)
+ *  - Model's input info, which is a final input's info after preprocessing (InputInfo::model)
+ */
 public class Input extends OpenVINOCls {
 
+    /**
+     * InputInfo class pointer.
+     */
     private Node mNode;
 
     private long index;
 
-    public Input(Node node,int index) {
+    public Input(Node node,long index) {
         super("Input-" + index,null);
         this.mNode = node;
         this.index = index;
